@@ -7,10 +7,10 @@ final loginTheme = ThemeData(
   dividerColor: Colors.white,
   scaffoldBackgroundColor: Colors.white,
   brightness: Brightness.light,
-
-  textTheme: TextTheme(
+  textTheme: const TextTheme(
     headline2: TextStyle(fontSize: 12.0, color: Colors.white70),
     button: TextStyle(fontSize: 16, color: Colors.white), //Used
     caption: TextStyle(fontSize: 14, color: Colors.white), //Used
-  ), textSelectionTheme: TextSelectionThemeData(cursorColor: Colors.blue[900]),
+  ),
+  textSelectionTheme: TextSelectionThemeData(cursorColor: Colors.blue[900]),
 );
